system-fa-py/requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
pymysql 
pydantic-settings 
python-dotenv
alembic
cryptography
python-jose[cryptography] 
passlib[bcrypt]